refactor : 피드백 반영 리팩토링 진행
- 모델 네이밍 수정
  - Decodable한 객체들을 Response로 네이밍 수정
  - Codable한 객체들을 DTO로 네이밍 수정

모델 이동
- NoteStatus를 DomainShared로 이동

Tuist Dependency 수정
- SharedDomainInterface의 dependency에 Core추가

그외:
- 변경된 모델들 및 import 때문에 빌드가 안되는 문제 수정
