+++ Projects/Core/Network/Interface/Sources/DTO/ArtistResponse.swift
@@ -1,5 +1,5 @@
 //
-//  ArtistDTO.swift
+//  ArtistResponse.swift
 //  CoreNetworkInterface
 //
 //  Created by 황인우 on 7/14/24.
@@ -7,7 +7,7 @@
 
 import Foundation
 
-public struct ArtistDTO: Decodable {
+public struct ArtistResponse: Decodable {
     public let id: Int
     public let name: String
     public let imageUrl: String?
