+++ Projects/Core/Network/Interface/Sources/DTO/SongResponse.swift
@@ -1,5 +1,5 @@
 //
-//  SongDTO.swift
+//  SongResponse.swift
 //  CoreNetworkInterface
 //
 //  Created by 황인우 on 8/18/24.
@@ -7,9 +7,9 @@
 
 import Foundation
 
-public struct SongDTO: Decodable {
+public struct SongResponse: Decodable {
     public let id: Int
     public let name: String
     public let imageUrl: String
-    public let artist: ArtistDTO
+    public let artist: ArtistResponse
 }
